Add temperature variables, migrate invariant and temperate
Two variants of the variable, one for block context (targets the mined block's location), and one for entity (targets the entity's location)
Temperates formula for tool damage reduction has to be the most complicated formula so far. Might also be my first time using swap. Got all the way up to 4 in the stack depth, largest so far.