Rebalance flamestance and entangled slightly
Entangled had solid numbers for mining speed and drawspeed (both worth just a bit more than an upgrade), but knockback resistance was a bit too high (worth 1.5 defense slots).
New number is the same as the knockback resistance modifier, but said modifier notably is max level 1.

Flamestance was just a bit too strong for a modifier that always works in the nether; downside is a bit too easy to mitigate.
Changed the power ranges for all effects to have equal best and worst case, and lowered the strength of melee damage and protection slightly.
Its still conditionally worth 1 ability slot (armor), but ranged is just 1.5 upgrades, and melee just under 1.5 upgrades.