Add manually created unknown bowstring textures to supporess the warning
Ideally I would ditch these entirely as the tinted ones look nearly identical. But I haven't found a good way to avoid it warning of missing textures when we do that.
Has to do with how Mojang uses the texture getter fetching a texture to mark it missing, but we use the texture getter to check if it exists. I might just need to handle existance using our other system if thats still around; but right now its too much work for 20 textures.