Add command to generate recipe removals
Warmup for the cooler feature
Just something I commonly do in most packs I make, might as well automate it. And if some mod dev finds it useful, then good for them
They are very aggressive; the casting filter for instance is to remove ingot recipes from the furnace as its a thing I like to do, but it will also remove brick recipes, so have to delete a few extras before saving the results.