Revamp recipe remove command to also support config presets
Keep in mind this is entirely volentary. Recipes are only removed if the command to remove them is run.

Requires a pretty messy hack to get the ingot smelting removal working, though it is just datagen so its not the worst thing. Does need a mantle feature to make it less messy though.
As part of the presets, add my three commonly performed tinkers related removals: ingot smelting, vanilla tool crafting, and netherite tool crafting.

Again, removals are volentary, just happens to be something far easier to implement as a java command than in python.