Add command to remove a single recipe by ID
Debatably some of these could be in Mantle, the big thing preventing that is Mantle doesn't have a way to do the castable predicate for the ingots at this time.
Once thats added, also need to consider the fact that putting it in mantle requires adding a namespace to the presets, not sure i want to do that without a suggestion provider.